refactor(game): split 796-line spells page into focused files under 250 lines
Reduces app/game/spells/page.tsx from 796 lines to 52 by extracting:

- 11 presentation components into app/game/spells/_components/
  - SpellsView — composed body with derived memos
    (spellByTierAndType, armableTiles, threatRanked) + bulk-arm state
  - TierSection — one tier row (3 columns)
  - SpellCell — single (tier, type) cell that delegates to a column-
    specific subcomponent
  - spell-cell/{ProductionControls, DefenseControls, BulkArmPanel,
    OffenseHint} — column-specific UI
  - ActiveProductionList, ArmedTilesList, ReportLog — page sections
  - SpellsGates — loading/sign-in/enlist/caste-required screens

- 4 lib modules into app/game/spells/_lib/
  - use-spells-data — server state + initial cache hydrate
  - use-spell-actions — callApi + cast/arm handlers + busyId/reports
  - types — response types + OwnerSummary
  - constants — TIERS, TYPE_COLUMNS, TYPE_LABEL, RARITY_TEXT

The _components and _lib prefixes are Next.js private folders, so the new
structure does not introduce any new routes. page.tsx is now a thin gate
that picks one of {Loading, SignIn, Enlist, CasteRequired, SpellsView}
based on the data-hook output.

No behavior change. All 1581 tests pass; tsc clean; lint clean.
